requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.6
django-allauth==0.45.0
django-ckeditor==6.1.0
django-filter==2.4.0
django-js-asset==1.2.2
django-mysql==3.12.0
django-storages==1.11.1
Django-Verify-Email==1.0.7
django-widget-tweaks==1.4.8
gunicorn==20.1.0
idna==3.2
mysqlclient==2.0.3
oauthlib==3.1.1
Pillow==8.3.1
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
sqlparse==0.4.1
urllib3==1.26.6
whitenoise==5.3.0
